2.0.109.5

+++ @vue-modules/multilanguage/init/vue-init-languages.js
@@ -9,8 +9,17 @@ Object.keys(ß.PAGES).forEach(function(page) {
 // then for each language, do the processing
 Object.keys(language_object).forEach(function(lang) {
 
-    ß.modules_process(ß.vue_modules, build_vue_src);
+    ß.modules_process(ß.vue_modules, build_vue_src_priority);
+    ß.modules_process(ß.vue_modules, build_vue_src_standard);
 
+    function build_vue_src_priority(module, module_dir, priority) {
+       if (priority) build_vue_src(module, module_dir);
+    }
+
+    function build_vue_src_standard(module, module_dir, priority) {
+       if (!priority) build_vue_src(module, module_dir);
+    }
+  
     function build_vue_src(module, module_dir) {
         const path = module_dir + "/vue";
         const dest = ß.VAR + "/vue/" + lang + "/src";
@@ -18,14 +27,14 @@ Object.keys(language_object).forEach(function(lang) {
         ß.fs.traverse_path_process_files(path, function(file_path) {
             let source = path + file_path;
             let target = dest + file_path;
-            let ext = ß.path.extname(source);
+            let ext = ß.path.extname(source); 
           	// translate files 
             if (ext === ".vue" || ext === ".js") {
                 let data = ß.fs.readFileSync(source, 'utf-8');
                 let processed_data = ß.lib.multilanguage.process(lang, data);
               	let target_dir = ß.path.dirname(target);
               	ß.fs.mkdirpSync(target_dir);
-              	ß.fs.writeFileSync(target, processed_data);
+              	if (!ß.fs.existsSync(target)) ß.fs.writeFileSync(target, processed_data);
                 return;
             }
             // and use links for non translateable files
